+++ src/_gettsim/transfers/unterhaltsvors.py
@@ -1,20 +1,14 @@
 """This module provides functions to compute advance alimony payments
 (Unterhaltsvorschuss)."""
 
-from _gettsim.shared import policy_info
+import numpy as np
 
-aggregate_by_group_id_unterhaltsvors = {
-    "_unterhaltsvorschuss_eink_above_income_threshold_fg": {
-        "group_id_to_aggregate_by": "fg_id",
-        "source_col": "_unterhaltsvorschuss_eink_above_income_threshold",
-        "aggr": "any",
-    },
-}
+from _gettsim.shared import join_numpy, policy_info
 
 aggregate_by_p_id_unterhaltsvors = {
-    "_unterhaltsvors_anspruch_eltern_m": {
+    "unterhaltsvors_zahlbetrag_eltern_m": {
         "p_id_to_aggregate_by": "p_id_kindergeld_empf",
-        "source_col": "_unterhaltsvors_anspruch_pro_kind_m",
+        "source_col": "unterhaltsvors_m",
         "aggr": "sum",
     },
 }
@@ -22,11 +16,12 @@ aggregate_by_p_id_unterhaltsvors = {
 
 @policy_info(params_key_for_rounding="unterhaltsvors")
 def unterhaltsvors_m(
-    alleinerz: bool,
     kind_unterh_erhalt_m: float,
-    _unterhaltsvors_anspruch_eltern_m: float,
-) -> float:
-    """Advance alimony payments (Unterhaltsvorschuss).
+    _unterhaltsvors_anspruch_kind_m: float,
+    parent_alleinerz: bool,
+):
+    """Advance alimony payments (Unterhaltsvorschuss) on child level after deducting
+    alimonies.
 
     Single Parents get alimony payments for themselves and for their child from the ex
     partner. If the ex partner is not able to pay the child alimony, the government pays
@@ -44,24 +39,49 @@ def unterhaltsvors_m(
 
     Parameters
     ----------
-    alleinerz
-        See basic input variable :ref:`alleinerz <alleinerz>`.
     kind_unterh_erhalt_m
         See :func:`kind_unterh_erhalt_m`.
-    _unterhaltsvors_anspruch_eltern_m
-        See :func:`_unterhaltsvors_anspruch_eltern_m`.
+    _unterhaltsvors_anspruch_kind_m
+        See :func:`_unterhaltsvors_anspruch_kind_m`.
+    parent_alleinerz
+        See :fung:`parent_alleinerz`.
 
     Returns
     -------
 
     """
-
-    if alleinerz:
-        out = max(_unterhaltsvors_anspruch_eltern_m - kind_unterh_erhalt_m, 0.0)
+    if parent_alleinerz:
+        out = max(_unterhaltsvors_anspruch_kind_m - kind_unterh_erhalt_m, 0.0)
     else:
         out = 0.0
 
     return out
+
+
+@policy_info(skip_vectorization=True)
+def parent_alleinerz(
+    p_id_kindergeld_empf: np.ndarray[int],
+    p_id: np.ndarray[int],
+    alleinerz: np.ndarray[bool],
+):
+    """Check if parent that receives Unterhaltsvorschuss is a single parent.
+
+    Only single parents receive Unterhaltsvorschuss.
+
+    Parameters
+    ----------
+    p_id_kindergeld_empf
+        See basic input variable :ref:`p_id_kindergeld_empf`.
+    p_id
+        See basic input variable :ref:`p_id`.
+    alleinerz
+        See basic input variable :ref:`alleinerz`.
+
+    Returns
+    -------
+
+    """
+    return join_numpy(p_id_kindergeld_empf, p_id, alleinerz)
 
 
 @policy_info(start_date="2023-01-01", name_in_dag="_kindergeld_erstes_kind_m")
@@ -108,21 +128,21 @@ def _kindergeld_erstes_kind_gestaffelt_m(
     return kindergeld_params["kindergeld"][1]
 
 
-def _unterhaltsvors_anspruch_pro_kind_m(
+def _unterhaltsvors_anspruch_kind_m(
     alter: int,
-    _unterhaltsvorschuss_eink_above_income_threshold_fg: bool,
+    _unterhaltsvorschuss_empf_eink_above_income_threshold: bool,
     _kindergeld_erstes_kind_m: float,
     unterhalt_params: dict,
     unterhaltsvors_params: dict,
 ) -> float:
-    """Claim for advance on alimony payment (Unterhaltsvorschuss) per child.
+    """Claim for advance on alimony payment (Unterhaltsvorschuss) on child level.
 
     Parameters
     ----------
     alter
         See basic input variable :ref:`alter <alter>`.
-    _unterhaltsvorschuss_eink_above_income_threshold_fg
-        See :func:`_unterhaltsvorschuss_eink_above_income_threshold_fg`.
+    _unterhaltsvorschuss_empf_eink_above_income_threshold
+        See :func:`_unterhaltsvorschuss_empf_eink_above_income_threshold`.
     _kindergeld_erstes_kind_m
         See :func:`_kindergeld_erstes_kind_m`.
     unterhalt_params
@@ -150,11 +170,37 @@ def _unterhaltsvors_anspruch_pro_kind_m(
     if (
         out > 0
         and (alter >= unterhaltsvors_params["altersgrenze_mindesteinkommen"])
-        and (not _unterhaltsvorschuss_eink_above_income_threshold_fg)
+        and (not _unterhaltsvorschuss_empf_eink_above_income_threshold)
     ):
         out = 0.0
 
     return out
+
+
+@policy_info(skip_vectorization=True)
+def _unterhaltsvorschuss_empf_eink_above_income_threshold(
+    p_id_kindergeld_empf: np.ndarray[int],
+    p_id: np.ndarray[int],
+    _unterhaltsvorschuss_eink_above_income_threshold: np.ndarray[bool],
+) -> np.ndarray[bool]:
+    """Income of Unterhaltsvorschuss recipient above threshold (this variable is
+    defined on child level).
+
+    Parameters
+    ----------
+    p_id_kindergeld_empf
+        See basic input variable :ref:`p_id_kindergeld_empf`.
+    p_id
+        See basic input variable :ref:`p_id`.
+    _unterhaltsvorschuss_eink_above_income_threshold
+        See :func:`_unterhaltsvorschuss_eink_above_income_threshold`.
+
+    Returns
+    -------
+    """
+    return join_numpy(
+        p_id_kindergeld_empf, p_id, _unterhaltsvorschuss_eink_above_income_threshold
+    )
 
 
 def _unterhaltsvorschuss_eink_above_income_threshold(